+++ resources/release-notes.md
@@ -5,5 +5,6 @@
 
 # Software
 
+- Add `mag.graphics.factory.LimitDisplayFactory` to limit the number of data points displayed in a figure
 - Fix HK data not being shown after long gaps
 - Fix `mag.Science/CompressionWidth` converted to `logical` instead of `double`